Add files via upload
This file was not used in the final capstone project for Cognizant but I created the code prior to learning the model in use required payment to use it. I added it to show use of secured API keys, error handling with a newer model, and the ability to use AI from different sources.